requirements.txt
psutil>=5.9.0
pyinstaller==6.3.0
pystray>=0.19.0
Pillow>=10.0.0
cryptography>=41.0.0
influxdb-client>=1.36.0
tkinter  # Inclus dans Python standard
